Add bug details, full shapefile, and nonsense
This commit adds more details about the bug to the README, including
steps to reproduce. I also realized the fixed shapefile from QGIS was
exporting just fine all along, but I was zipping it into a directory
and therefore was just entering the path wrong. So obviously... ogr
didn't recognize that. There is no issue with QGIS